+++ src/components/Carousel/Carousel.js
@@ -1,9 +1,10 @@
 import AliceCarousel from "react-alice-carousel";
 import "react-alice-carousel/lib/alice-carousel.css";
-import image1 from "../img/image1.jpg";
-import image2 from "../img/image2.jpg";
-import image3 from "../img/image3.jpg";
-import image4 from "../img/image4.jpg";
+import image1 from "./img/image1.jpg";
+import image2 from "./img/image2.jpg";
+import image3 from "./img/image3.jpg";
+import image4 from "./img/image4.jpg";
+import "./carousel.css";
 
 function Carousel() {
   return (
