+++ src/components/Carousel/carousel.css
@@ -1,0 +1,6 @@
+.carousel {
+  background-color: white;
+  position: flex;
+  width: 100%;
+  margin-top: 0;
+}
