+++ src/components/Carousel/img/image2.jpg
