+++ src/index.js
@@ -1,7 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import App from "./App";
-import "./style.css";
 
 ReactDOM.render(
   <React.StrictMode>
