--- src/pages/home/Navbar.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Navbar = () => {
-  return <div>soy un navbar</div>;
-};
-
-export default Navbar;
