+++ src/pages/home/homeStyles.css
@@ -1,0 +1,29 @@
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+}
+
+.searchBar {
+  width: 100%;
+  display: flex;
+  align-items: center;
+}
+
+.cover {
+  display: flex;
+}
+
+.sliderimg {
+  width: 100%;
+  height: 500px;
+  object-fit: cover;
+}
+
+.App {
+  color: black;
+}
+
+.newUser {
+  background-color: orangered;
+}
