+++ src/pages/home/index.js
@@ -1,31 +1,24 @@
 import React from "react";
 
-import Carousel from "../../components/Carousel";
-import Search from "../../components/Search";
-import logo from "../../img/Logo.jpg";
-import Navbar from "./Navbar";
+import Carousel from "../../components/Carousel/Carousel";
+import Search from "../../components/Search/Search";
+import Navbar from "./Navbar/Navbar";
+import CoverInfo from "../../components/Cover/CoverInfo";
+import "./homeStyles.css";
 
 const Home = () => {
   return (
     <div className='App'>
-      <Navbar />
-      <a className='logotext' href='/home'>
-        <img className='logo' src={logo} alt='logo' />
-        Soundmist
-      </a>
-      <div className='directions'>
-        <a className='creators' href='/Creators'>
-          For Creators
-        </a>
-        <a className='newUser' href='/NewUser'>
-          Create Account
-        </a>
-        <a className='login' href='/SignIn'>
-          Sign in
-        </a>
+      <div>
+        <Navbar />
+        <Carousel />
       </div>
-      <Carousel />
-      <Search />
+      <div className='searchBar'>
+        <Search />
+      </div>
+      <div className='cover'>
+        <CoverInfo />
+      </div>
     </div>
   );
 };
